+++ internal/js_parser/js_parser.go
@@ -13538,36 +13538,6 @@ func (p *parser) scanForImportsAndExports(stmts []js_ast.Stmt) (result importsEx
 
 			if p.options.mode != config.ModePassThrough {
 				if s.StarNameLoc != nil {
-					// If we're bundling a star import and the namespace is only ever
-					// used for property accesses, then convert each unique property to
-					// a clause item in the import statement and remove the star import.
-					// That will cause the bundler to bundle them more efficiently when
-					// both this module and the imported module are in the same group.
-					//
-					// Before:
-					//
-					//   import * as ns from 'foo'
-					//   console.log(ns.a, ns.b)
-					//
-					// After:
-					//
-					//   import {a, b} from 'foo'
-					//   console.log(a, b)
-					//
-					// This is not done if the namespace itself is used, because in that
-					// case the code for the namespace will have to be generated. This is
-					// determined by the symbol count because the parser only counts the
-					// star import as used if it was used for something other than a
-					// property access:
-					//
-					//   import * as ns from 'foo'
-					//   console.log(ns, ns.a, ns.b)
-					//
-					convertStarToClause := p.symbols[s.NamespaceRef.InnerIndex].UseCountEstimate == 0
-					if convertStarToClause && !keepUnusedImports {
-						s.StarNameLoc = nil
-					}
-
 					// "importItemsForNamespace" has property accesses off the namespace
 					if importItems, ok := p.importItemsForNamespace[s.NamespaceRef]; ok && len(importItems) > 0 {
 						// Sort keys for determinism
@@ -13577,52 +13547,32 @@ func (p *parser) scanForImportsAndExports(stmts []js_ast.Stmt) (result importsEx
 						}
 						sort.Strings(sorted)
 
-						if convertStarToClause {
-							// Create an import clause for these items. Named imports will be
-							// automatically created later on since there is now a clause.
-							items := make([]js_ast.ClauseItem, 0, len(importItems))
-							for _, alias := range sorted {
-								name := importItems[alias]
-								originalName := p.symbols[name.Ref.InnerIndex].OriginalName
-								items = append(items, js_ast.ClauseItem{
-									Alias:        alias,
-									AliasLoc:     name.Loc,
-									Name:         name,
-									OriginalName: originalName,
-								})
-								p.declaredSymbols = append(p.declaredSymbols, js_ast.DeclaredSymbol{
-									Ref:        name.Ref,
-									IsTopLevel: true,
-								})
+						// Create named imports for these property accesses. This will
+						// cause missing imports to generate useful warnings.
+						//
+						// It will also improve bundling efficiency for internal imports
+						// by still converting property accesses off the namespace into
+						// bare identifiers even if the namespace is still needed.
+						for _, alias := range sorted {
+							name := importItems[alias]
+							p.namedImports[name.Ref] = js_ast.NamedImport{
+								Alias:             alias,
+								AliasLoc:          name.Loc,
+								NamespaceRef:      s.NamespaceRef,
+								ImportRecordIndex: s.ImportRecordIndex,
 							}
-							if s.Items != nil {
-								// The syntax "import {x}, * as y from 'path'" isn't valid
-								panic("Internal error")
-							}
-							s.Items = &items
-						} else {
-							// If we aren't converting this star import to a clause, still
-							// create named imports for these property accesses. This will
-							// cause missing imports to generate useful warnings.
-							//
-							// It will also improve bundling efficiency for internal imports
-							// by still converting property accesses off the namespace into
-							// bare identifiers even if the namespace is still needed.
-							for _, alias := range sorted {
-								name := importItems[alias]
-								p.namedImports[name.Ref] = js_ast.NamedImport{
-									Alias:             alias,
-									AliasLoc:          name.Loc,
-									NamespaceRef:      s.NamespaceRef,
-									ImportRecordIndex: s.ImportRecordIndex,
-								}
 
-								// Make sure the printer prints this as a property access
-								p.symbols[name.Ref.InnerIndex].NamespaceAlias = &js_ast.NamespaceAlias{
-									NamespaceRef: s.NamespaceRef,
-									Alias:        alias,
-								}
+							// Make sure the printer prints this as a property access
+							p.symbols[name.Ref.InnerIndex].NamespaceAlias = &js_ast.NamespaceAlias{
+								NamespaceRef: s.NamespaceRef,
+								Alias:        alias,
 							}
+
+							// Also record these automatically-generated top-level namespace alias symbols
+							p.declaredSymbols = append(p.declaredSymbols, js_ast.DeclaredSymbol{
+								Ref:        name.Ref,
+								IsTopLevel: true,
+							})
 						}
 					}
 				}