+++ scripts/end-to-end-tests.js
@@ -1480,7 +1480,7 @@
       `,
       'node_modules/pkg/index.mjs': ``,
     }, {
-      expectedStderr: `▲ [WARNING] Import "default" will always be undefined because there is no matching export
+      expectedStderr: `▲ [WARNING] Import "default" will always be undefined because there is no matching export in "node_modules/pkg/index.mjs"
 
     in.js:3:15:
       3 │         if (ns.default !== void 0) throw 'fail'
@@ -1495,7 +1495,7 @@
       `,
       'node_modules/pkg/index.mts': ``,
     }, {
-      expectedStderr: `▲ [WARNING] Import "default" will always be undefined because there is no matching export
+      expectedStderr: `▲ [WARNING] Import "default" will always be undefined because there is no matching export in "node_modules/pkg/index.mts"
 
     in.js:3:15:
       3 │         if (ns.default !== void 0) throw 'fail'
@@ -1523,13 +1523,27 @@
       }`,
       'node_modules/pkg/index.js': ``,
     }, {
-      expectedStderr: `▲ [WARNING] Import "default" will always be undefined because there is no matching export
+      expectedStderr: `▲ [WARNING] Import "default" will always be undefined because there is no matching export in "node_modules/pkg/index.js"
 
     in.js:3:15:
       3 │         if (ns.default !== void 0) throw 'fail'
         ╵                ~~~~~~~
 
 `,
+    }),
+    test(['in.js', '--outfile=node.js', '--bundle', '--external:pkg'], {
+      'in.js': `
+        import * as ns from 'pkg'
+        if (ns.default === void 0) throw 'fail'
+      `,
+      'node_modules/pkg/index.js': ``,
+    }),
+    test(['in.js', '--outfile=node.js', '--bundle', '--external:pkg'], {
+      'in.js': `
+        import * as ns from 'pkg'
+        if (ns.foo !== void 0) throw 'fail'
+      `,
+      'node_modules/pkg/index.js': ``,
     }),
   )